+++ backend/repository/component_repository_json_impl.go
@@ -57,7 +57,7 @@ func NewJSONComponentRepository(fr utils.FileReader) *JSONComponentRepository {
 }
 
 func (r *JSONComponentRepository) FindByFilePath(path string) (entities.Component, error) {
-	resultFilePath := config.GetInstance().ResultFilePath
+	resultFilePath := config.GetInstance().GetResultFilePath()
 
 	resultFileBytes, err := r.fr.ReadFile(resultFilePath)
 	if err != nil {