+++ backend/repository/scanoss_settings_repository_json_impl.go
@@ -47,35 +47,63 @@ func NewScanossSettingsJsonRepository(fr utils.FileReader) ScanossSettingsReposi
 }
 
 func (r *ScanossSettingsJsonRepository) Init() error {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
+	cfg := config.GetInstance()
 
-	sf, err := r.Read()
-	if err != nil {
-		log.Fatal().Err(err).Msgf("Error reading settings file: %v", config.GetInstance().ScanSettingsFilePath)
+	if cfg == nil {
+		err := fmt.Errorf("config is not initialized")
+		log.Error().Err(err).Msg("Error initializing ScanossSettingsJsonRepository")
 		return err
 	}
 
-	entities.ScanossSettingsJson = &entities.ScanossSettings{
-		SettingsFile: &sf,
-	}
+	r.setSettingsFile(cfg.GetScanSettingsFilePath())
+
+	cfg.RegisterListener(r.onConfigChange)
 
 	return nil
 }
 
+// Triggered when the config is changed (e.g. scan root, scan settings file path, etc.)
+func (r *ScanossSettingsJsonRepository) onConfigChange(newCfg *config.Config) {
+	r.setSettingsFile(newCfg.GetScanSettingsFilePath())
+}
+
+func (r *ScanossSettingsJsonRepository) setSettingsFile(path string) {
+	sf, err := r.Read()
+
+	if err != nil {
+		log.Error().Err(err).Msgf("Error reading settings file: %v", path)
+		return
+	}
+
+	if entities.ScanossSettingsJson == nil {
+		entities.ScanossSettingsJson = &entities.ScanossSettings{
+			SettingsFile: &sf,
+		}
+	}
+
+	entities.ScanossSettingsJson.SettingsFile = &sf
+}
+
 func (r *ScanossSettingsJsonRepository) Save() error {
+	r.mutex.Lock()
+	defer r.mutex.Unlock()
+
 	sf := r.GetSettings()
-	if err := utils.WriteJsonFile(config.GetInstance().ScanSettingsFilePath, sf); err != nil {
+	if err := utils.WriteJsonFile(config.GetInstance().GetScanSettingsFilePath(), sf); err != nil {
 		return err
 	}
 	return nil
 }
 
 func (r *ScanossSettingsJsonRepository) Read() (entities.SettingsFile, error) {
-	if config.GetInstance() == nil {
+	r.mutex.RLock()
+	defer r.mutex.RUnlock()
+
+	cfg := config.GetInstance()
+	if cfg == nil {
 		return entities.SettingsFile{}, fmt.Errorf("config is not initialized")
 	}
-	scanSettingsFileBytes, err := r.fr.ReadFile(config.GetInstance().ScanSettingsFilePath)
+	scanSettingsFileBytes, err := r.fr.ReadFile(cfg.GetScanSettingsFilePath())
 	if err != nil {
 		if errors.Is(err, os.ErrNotExist) {
 			return entities.SettingsFile{}, nil