+++ backend/service/scan_service_python_cli_impl.go
@@ -30,7 +30,10 @@ import (
 	"io"
 	"os"
 	"os/exec"
+	"path/filepath"
 
+	"github.com/rs/zerolog/log"
+	"github.com/scanoss/scanoss.cc/backend/entities"
 	"github.com/scanoss/scanoss.cc/internal/config"
 	"github.com/wailsapp/wails/v2/pkg/runtime"
 )
@@ -112,6 +115,9 @@ func (s *ScanServicePythonImpl) executeScanWithPipes(args []string) (*exec.Cmd, 
 
 	cmdArgs = append(cmdArgs, sensitiveArgs...)
 
+	// If the output folder does not exist, create it. This should be handled by the python cli
+	s.maybeCreateOutputFolder(args)
+
 	cmd := exec.Command(s.cmd, cmdArgs...)
 
 	stdout, err := cmd.StdoutPipe()
@@ -182,12 +188,12 @@ func (s *ScanServicePythonImpl) GetDefaultScanArgs() []string {
 	args := []string{}
 	cfg := config.GetInstance()
 
-	if cfg.ResultFilePath != "" {
-		args = append(args, "--output", cfg.ResultFilePath)
+	if cfg.GetResultFilePath() != "" {
+		args = append(args, "--output", cfg.GetResultFilePath())
 	}
 
-	if cfg.ScanSettingsFilePath != "" {
-		args = append(args, "--settings", cfg.ScanSettingsFilePath)
+	if cfg.GetScanSettingsFilePath() != "" {
+		args = append(args, "--settings", cfg.GetScanSettingsFilePath())
 	}
 
 	return args
@@ -197,12 +203,12 @@ func (s *ScanServicePythonImpl) GetSensitiveDefaultScanArgs() []string {
 	args := make([]string, 0)
 	cfg := config.GetInstance()
 
-	if cfg.ApiToken != "" {
-		args = append(args, "--key", cfg.ApiToken)
+	if cfg.GetApiToken() != "" {
+		args = append(args, "--key", cfg.GetApiToken())
 	}
 
-	if cfg.ApiUrl != "" {
-		args = append(args, "--apiurl", fmt.Sprintf("%s/scan/direct", cfg.ApiUrl))
+	if cfg.GetApiUrl() != "" {
+		args = append(args, "--apiurl", fmt.Sprintf("%s/scan/direct", cfg.GetApiUrl()))
 	}
 
 	return args
@@ -212,4 +218,30 @@ func (s *ScanServicePythonImpl) emitEvent(eventName string, data ...interface{})
 	if s.ctx != nil {
 		runtime.EventsEmit(s.ctx, eventName, data...)
 	}
+}
+
+func (s *ScanServicePythonImpl) GetScanArgs() []entities.ScanArgDef {
+	return entities.ScanArguments
+}
+
+func (s *ScanServicePythonImpl) maybeCreateOutputFolder(args []string) {
+	outputPath := s.getOutputPathFromArgs(args)
+	outputFolder := filepath.Dir(outputPath)
+	if outputFolder != "" {
+		if _, err := os.Stat(outputFolder); os.IsNotExist(err) {
+			log.Info().Msgf("The provided output path does not exist. Creating it: %s", outputFolder)
+			if err := os.MkdirAll(outputFolder, os.ModePerm); err != nil {
+				log.Error().Err(err).Msgf("Failed to create output folder: %s", outputFolder)
+			}
+		}
+	}
+}
+
+func (s *ScanServicePythonImpl) getOutputPathFromArgs(args []string) string {
+	for i := 0; i < len(args)-1; i++ {
+		if args[i] == "--output" {
+			return args[i+1]
+		}
+	}
+	return ""
 }